#668 Added fix for math functions not being able to take other functions of expressions as input.
