+++ vuu/src/main/resources/runconfigurations/SimulMain.run.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="SimulMain" type="Application" factoryName="Application" folderName="runConfigs">
-    <option name="ALTERNATIVE_JRE_PATH" value="$USER_HOME$/.sdkman/candidates/java/11.0.8.j9-adpt" />
+    <option name="ALTERNATIVE_JRE_PATH" value="$USER_HOME$/.sdkman/candidates/java/16.0.2-open" />
     <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="true" />
     <option name="MAIN_CLASS_NAME" value="org.finos.vuu.SimulMain" />
     <module name="vuu" />
